+++ named/server_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/goph/serverz"
+	"github.com/goph/serverz/internal"
 	"github.com/goph/serverz/internal/mocks"
 	"github.com/goph/serverz/named"
 	"github.com/stretchr/testify/assert"
@@ -15,7 +16,7 @@ func TestNewServer(t *testing.T) {
 	server := named.NewServer(s, "name")
 
 	assert.Equal(t, s, server.Server)
-	assert.Equal(t, "name", server.ServerName)
+	assert.Equal(t, "name", server.Name)
 }
 
 func TestServerIsAServer(t *testing.T) {
@@ -24,8 +25,9 @@ func TestServerIsAServer(t *testing.T) {
 
 func TestServer_Name(t *testing.T) {
 	server := &named.Server{
-		ServerName: "name",
+		Name: "name",
 	}
 
-	assert.Equal(t, "name", server.Name())
+	assert.Implements(t, (*internal.NamedServer)(nil), server)
+	assert.Equal(t, "name", server.GetName())
 }